Bugfixing der aufgenommenen Bugs auf itsLearning

+++ src/main/java/datastorage/DAO.java
@@ -6,11 +6,11 @@ import java.util.List;
 public interface DAO<T> {
     void create(T t) throws SQLException;
 
-    T read(int key) throws SQLException;
+    T read(long key) throws SQLException;
 
     List<T> readAll() throws SQLException;
 
     void update(T t) throws SQLException;
 
-    void deleteById(int key) throws SQLException;
+    void deleteById(long key) throws SQLException;
 }
